[Decomp][Unit] Cohesion split: extract speed + movement-state methods to UnitSpeed.cpp
